+++ CHANGELOG.markdown
@@ -1,6 +1,9 @@
 ## Pending
 - Add Prism AutoInstruments Support (#582) (#587)
 - Add HTTPX instrumentation (#588)
+- Add ability to automatically capture Sidekiq job args as context
+  - `job_params_capture` - Set to true to enable job argument capturing
+  - `job_params_filter`  - A list of arguments to filter (automatically includes Rails filtered_parameters) 
 - Fix user error context being incorrectly flattened (#581)
 - Handle Delayed Job PerformableMethod jobs for error tracking (#584)
 - Require 'httpclient' library on instrumentation install (#586)